feat(providers): rename gemini provider to google provider and support agy CLI auth
